Move window operators and tests to windowing package
The api package is also called windowing, this harmonizes the package
names.
