+++ flink-staging/flink-streaming/flink-streaming-core/src/main/java/org/apache/flink/streaming/runtime/operators/windowing/package-info.java
@@ -19,4 +19,4 @@
  * This package contains the operators that implement the various window operations
  * on data streams. 
  */
-package org.apache.flink.streaming.runtime.operators.windows;
+package org.apache.flink.streaming.runtime.operators.windowing;
